crawler/requirements.txt
arxiv==0.2.2
git+https://github.com/titipata/arxivpy
git+https://github.com/okfn/pdftables
certifi==2018.11.29
chardet==3.0.4
feedparser==5.2.1
idna==2.7
numpy==1.15.1
pandas==0.23.4
peewee==3.6.4
peewee-async==0.6.0a0
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.7
requests==2.20.1
six==1.11.0
tqdm==4.28.1
urllib3==1.24.1
pdfminer.six
pyPDF2